% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_eigengene_explained_variance.R
\name{test_eigengene_explained_variance}
\alias{test_eigengene_explained_variance}
\title{Test the significance of a target eigengene}
\usage{
test_eigengene_explained_variance(OriginalData, Original_blockwiseModules,
  target_module, permutations = 999, bicor = FALSE, ...)
}
\arguments{
\item{OriginalData}{Matrix or data frame
containing the original data on which the blockwiseModules function has been run
(observations in rows, variables in columns).}

\item{Original_blockwiseModules}{output of the blockwiseModules function}

\item{target_module}{module whose largest eigenvalue/eigengene will be tested}

\item{permutations}{number of permutations to use}

\item{bicor}{whether one should use the standard eigengene function in WGCNA (default), or bicor}

\item{...}{further parameters to be passed to either moduleEigengenes or bicor}
}
\value{
The function outputs a list with the following elements:
 \describe{
  \item{moduleEigengenes_ALL}{The eigengenes for all the modules (will be empty if bicor=TRUE)}
  \item{explained_variance_target_module}{The observed explained variance for the target module}
  \item{p_value}{The p value obtained through permutation}
}
}
\description{
Performs a permutation test of significance for the eigengene
}
\details{
Given a dataset, a set of results from blockwiseModules, and a target module,
this function performs a permutation test of the variance explained by the eigengene
(first eigenvalue of the correlation matrix of the genes belonging to the module).
The target module should be expressed in the same way as when using blockwiseModules
(i.e., color or number)
Notice that option bicor=TRUE can be extremely computationally demanding
}
\section{Citation}{

If you use this function please cite Fruciano et al. in press
}

\references{
Fruciano C., Franchini P., Meyer A. in press. Divergent Allometric Trajectories in Gene Expression and Coexpression Produce Species Differences in Sympatrically SpeciatingMidas Cichlid Fish. Genome Biology and Evolution
}
