% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_RV_target_module.R
\name{test_RV_target_module}
\alias{test_RV_target_module}
\title{Analysis of the association between a target module and one or more variables}
\usage{
test_RV_target_module(OriginalData, Original_blockwiseModules,
  target_module, X, permutations = 999, rarefied = FALSE,
  reps_rarefaction = 1000, samplesize_rarefaction = NULL)
}
\arguments{
\item{OriginalData}{Matrix or data frame
containing the original data on which the blockwiseModules function has been run
(observations in rows, variables in columns).}

\item{Original_blockwiseModules}{output of the blockwiseModules function}

\item{target_module}{module whose association will be assessed}

\item{X}{one or more variables whose association with the target module will be tested}

\item{permutations}{number of permutations to use in the test of association}

\item{rarefied}{if TRUE, it also performs rarefaction-based estimation of Escoufier RV}

\item{reps_rarefaction}{number of replicated rarefied samples}

\item{samplesize_rarefaction}{sample size to which rarefy (required if rarefied=TRUE)}
}
\value{
The function outputs a list with the following elements:
 \describe{
  \item{ObsRV}{Observed RV coefficient}
  \item{p_value_perm}{p value obtained through permutation}
  \item{Rarefied_RV}{Results of the rarefaction analysis (empty if rarefied=FALSE)}
}
}
\description{
Given a dataset, a set of results from blockwiseModules, a target module, and one or more variables,
this function performs a permutation test based on Escoufier RV.
Optionally, it can also compute rarefied estimates of RV.
Notice that this function requires the package GeometricMorphometricsMix
and that the analysis can be computationally demanding for large modules
}
